# API_Extended_Document/doc/OpenAPI2/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "API used to manage the extended documents in UDV Server.",
    "version" : "1.0.0",
    "title" : "UDV-Server API"
  },
  "host" : "0.0.0.0:5000",
  "basePath" : "/",
...
  "schemes" : [ "https" ],
  "paths" : {
    "/login" : {
      "post" : {
        "tags" : [ "Users" ],
        "summary" : "Log in and retrieve a token",
        "description" : "Authenticate a user by giving them a auth token",
        "consumes" : [ "multipart/form-data" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "username",
          "in" : "formData",
          "description" : "Username of the user",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "password",
          "in" : "formData",
          "description" : "Password of the user",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Successfully logged in",
            "schema" : {
              "$ref" : "#/definitions/TokenItem"
            }
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "500" : {
            "description" : "Unexpected error"
          }
        }
      }
    },
    "/user" : {
      "post" : {
        "tags" : [ "Users" ],
        "summary" : "Create a new user",
        "description" : "Add a new User by passing informations below\n",
        "consumes" : [ "multipart/form-data" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "lastName",
          "in" : "formData",
          "description" : "Pass the lastName of user you want to create",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "firstName",
          "in" : "formData",
          "description" : "Pass the firstName of user you want to create",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "email",
          "in" : "formData",
          "description" : "Pass the email of user you want to create",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "username",
          "in" : "formData",
          "description" : "Pass the username of user you want to create",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "password",
          "in" : "formData",
          "description" : "Pass the password of user you want to create",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "201" : {
            "description" : "Information about the new user",
            "schema" : {
              "$ref" : "#/definitions/UserItem"
            }
          },
          "422" : {
            "description" : "Integrity error"
          },
          "500" : {
            "description" : "Unexpected error"
          }
        }
      }
    },
    "/user/me" : {
      "get" : {
        "tags" : [ "Users" ],
        "summary" : "Retrieve the connected user",
        "description" : "Retrieve the logged in user with the given token\n",
        "produces" : [ "application/json" ],
        "parameters" : [ ],
        "responses" : {
          "200" : {
            "description" : "The connected user's information",
            "schema" : {
              "$ref" : "#/definitions/UserItem"
            }
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "500" : {
            "description" : "Unexpected error"
          }
        },
        "security" : [ {
          "Bearer" : [ ]
        } ]
      }
    },
    "/user/{id}" : {
      "get" : {
        "tags" : [ "Users" ],
        "summary" : "Retrieve user by their ID",
        "description" : "Retrieve a specific User by passing the id\n",
        "consumes" : [ "multipart/form-data" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "Pass the id of user you want to retrieve",
          "required" : true,
          "type" : "integer"
        } ],
        "responses" : {
          "200" : {
            "description" : "The user's information",
            "schema" : {
              "$ref" : "#/definitions/UserItem"
            }
          },
          "404" : {
            "description" : "Not found"
          },
          "500" : {
            "description" : "Unexpected error"
          }
        }
      }
    },
    "/user/grant" : {
      "post" : {
        "tags" : [ "Users" ],
        "summary" : "Create a privileged user",
        "description" : "Create a user with the specific role",
        "consumes" : [ "multipart/form-data" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "lastName",
          "in" : "formData",
          "description" : "Pass the lastName of user you want to create",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "firstName",
          "in" : "formData",
          "description" : "Pass the firstName of user you want to create",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "email",
          "in" : "formData",
          "description" : "Pass the email of user you want to create",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "username",
          "in" : "formData",
          "description" : "Pass the username of user you want to create",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "password",
          "in" : "formData",
          "description" : "Pass the password of user you want to create",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "role",
          "in" : "formData",
          "description" : "The role of the user",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "The user's information",
            "schema" : {
              "$ref" : "#/definitions/UserItem"
            }
          },
          "404" : {
            "description" : "Not found"
          },
          "500" : {
            "description" : "Unexpected error"
          }
        },
        "security" : [ {
          "Bearer" : [ ]
        } ]
      }
    },
    "/document" : {
      "get" : {
        "tags" : [ "Documents" ],
        "summary" : "Get all extended documents",
        "parameters" : [ ],
        "responses" : {
          "200" : {
            "description" : "Successfully get the extended documents",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/ExtendedDocument"
              }
            }
          },
          "500" : {
            "description" : "Unexpected error"
          }
        }
      },
      "post" : {
        "tags" : [ "Documents" ],
        "summary" : "Create a new extended document",
        "consumes" : [ "multipart/form-data" ],
        "parameters" : [ {
          "name" : "title",
          "in" : "formData",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "subject",
          "in" : "formData",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "description",
          "in" : "formData",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "refDate",
          "in" : "formData",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "publicationDate",
          "in" : "formData",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "type",
          "in" : "formData",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "quaternionX",
          "in" : "formData",
          "required" : true,
          "type" : "number"
        }, {
          "name" : "quaternionY",
          "in" : "formData",
          "required" : true,
          "type" : "number"
        }, {
          "name" : "quaternionZ",
          "in" : "formData",
          "required" : true,
          "type" : "number"
        }, {
          "name" : "quaternionW",
          "in" : "formData",
          "required" : true,
          "type" : "number"
        }, {
          "name" : "positionX",
          "in" : "formData",
          "required" : true,
          "type" : "number"
        }, {
          "name" : "positionY",
          "in" : "formData",
          "required" : true,
          "type" : "number"
        }, {
          "name" : "positionZ",
          "in" : "formData",
          "required" : true,
          "type" : "number"
        }, {
          "name" : "file",
          "in" : "formData",
          "required" : true,
          "type" : "file"
        } ],
        "responses" : {
          "201" : {
            "description" : "Successfully created an extended document",
            "schema" : {
              "$ref" : "#/definitions/ExtendedDocument"
            }
          },
          "400" : {
            "description" : "Bad request"
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "422" : {
            "description" : "Integrity error"
          },
          "500" : {
            "description" : "Unexpected error"
          }
        },
        "security" : [ {
          "Bearer" : [ ]
        } ]
      }
    },
    "/document/{id}" : {
      "get" : {
        "tags" : [ "Documents" ],
        "summary" : "Get an extended document with a specific id",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "ID of the extendedDoc.",
          "required" : true,
          "type" : "integer"
        } ],
        "responses" : {
          "200" : {
            "description" : "Successfully get the extendedDoc",
            "schema" : {
              "$ref" : "#/definitions/ExtendedDocument"
            }
          },
          "404" : {
            "description" : "Not found"
          },
          "500" : {
            "description" : "Unexpected error"
          }
        }
      },
      "put" : {
        "tags" : [ "Documents" ],
        "summary" : "Update an extended document",
        "consumes" : [ "multipart/form-data" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "ID of the document.",
          "required" : true,
          "type" : "integer"
        }, {
          "name" : "title",
          "in" : "formData",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "subject",
          "in" : "formData",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "description",
          "in" : "formData",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "refDate",
          "in" : "formData",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "publicationDate",
          "in" : "formData",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "type",
          "in" : "formData",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "quaternionX",
          "in" : "formData",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "quaternionY",
          "in" : "formData",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "quaternionZ",
          "in" : "formData",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "quaternionW",
          "in" : "formData",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "positionX",
          "in" : "formData",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "positionY",
          "in" : "formData",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "positionZ",
          "in" : "formData",
          "required" : false,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Successfully updated the document",
            "schema" : {
              "$ref" : "#/definitions/ExtendedDocument"
            }
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "404" : {
            "description" : "Not found"
          },
          "422" : {
            "description" : "Integrity error"
          },
          "500" : {
            "description" : "Unexpected error"
          }
        },
        "security" : [ {
          "Bearer" : [ ]
        } ]
      },
      "delete" : {
        "tags" : [ "Documents" ],
        "summary" : "Delete an extended document with a specific id",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "ID of the document.",
          "required" : true,
          "type" : "integer"
        } ],
        "responses" : {
          "200" : {
            "description" : "Successfully deleted the document",
            "schema" : {
              "$ref" : "#/definitions/ExtendedDocument"
            }
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "404" : {
            "description" : "Not found"
          },
          "500" : {
            "description" : "Unexpected error"
          }
        },
        "security" : [ {
          "Bearer" : [ ]
        } ]
      }
    },
    "/document/{id}/file" : {
      "get" : {
        "tags" : [ "Document files" ],
        "summary" : "Get the file associated to a document",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "ID of the extendedDoc.",
          "required" : true,
          "type" : "integer"
        } ],
        "responses" : {
          "200" : {
            "description" : "Successfully get the document's image with the specified id\n"
          },
          "404" : {
            "description" : "Not found"
          },
          "500" : {
            "description" : "Unexpected error"
          }
        }
      },
      "post" : {
        "tags" : [ "Document files" ],
        "summary" : "Upload and associate a file to document",
        "consumes" : [ "multipart/form-data" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "ID of the document",
          "required" : true,
          "type" : "integer"
        }, {
          "name" : "file",
          "in" : "formData",
          "description" : "The file to upload",
          "required" : true,
          "type" : "file"
        } ],
        "responses" : {
          "200" : {
            "description" : "Successfully uploaded the document's file"
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "404" : {
            "description" : "Not found"
          },
          "422" : {
            "description" : "Integrity error"
          },
          "500" : {
            "description" : "Unexpected error"
          }
        },
        "security" : [ {
          "Bearer" : [ ]
        } ]
      },
      "delete" : {
        "tags" : [ "Document files" ],
        "summary" : "Delete the file associated to a document",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "ID of the extendedDoc.",
          "required" : true,
          "type" : "integer"
        } ],
        "responses" : {
          "204" : {
            "description" : "Successfully delete the extendedDocument file"
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "404" : {
            "description" : "Not found"
          },
          "500" : {
            "description" : "Unexpected error"
          }
        },
        "security" : [ {
          "Bearer" : [ ]
        } ]
      }
    },
    "/document/in_validation" : {
      "get" : {
        "tags" : [ "Document validation" ],
        "summary" : "Get the documents in validation",
        "parameters" : [ ],
        "responses" : {
          "200" : {
            "description" : "The retrieved documents",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/ExtendedDocument"
              }
            }
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "500" : {
            "description" : "Unexpected error"
          }
        },
        "security" : [ {
          "Bearer" : [ ]
        } ]
      }
    },
    "/document/validate" : {
      "post" : {
        "tags" : [ "Document validation" ],
        "summary" : "Validate a document",
        "consumes" : [ "multipart/form-data" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "formData",
          "description" : "ID of the extended document in validation",
          "required" : true,
          "type" : "integer"
        } ],
        "responses" : {
          "200" : {
            "description" : "Successfuly validated the document"
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "404" : {
            "description" : "Not found"
          },
          "500" : {
            "description" : "Unexpected error"
          }
        },
        "security" : [ {
          "Bearer" : [ ]
        } ]
      }
    },
    "/document/{id}/comment" : {
      "get" : {
        "tags" : [ "Document comments" ],
        "summary" : "Get all the comments associated to a document",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "ID of the extendedDoc.",
          "required" : true,
          "type" : "integer"
        } ],
        "responses" : {
          "200" : {
            "description" : "The retrieved comments for this document",
            "schema" : {
              "$ref" : "#/definitions/Comment"
            }
          },
          "404" : {
            "description" : "Not found"
          },
          "500" : {
            "description" : "Unexpected error"
          }
        }
      },
      "post" : {
        "tags" : [ "Document comments" ],
        "summary" : "Create a new comment, associated to a document",
        "consumes" : [ "multipart/form-data" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "ID of the extendedDoc.",
          "required" : true,
          "type" : "integer"
        }, {
          "name" : "description",
          "in" : "formData",
          "description" : "Comment",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "201" : {
            "description" : "Successfully create a new comment, associated to a document",
            "schema" : {
              "$ref" : "#/definitions/Comment"
            }
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "404" : {
            "description" : "Not found"
          },
          "422" : {
            "description" : "Integrity error"
          },
          "500" : {
            "description" : "Unexpected error"
          }
        },
        "security" : [ {
          "Bearer" : [ ]
        } ]
      }
    },
    "/comment/{id}" : {
      "get" : {
        "tags" : [ "Document comments" ],
        "summary" : "Get a specific comment",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "ID of the comment",
          "required" : true,
          "type" : "integer"
        } ],
        "responses" : {
          "200" : {
            "description" : "The retrieved comment",
            "schema" : {
              "$ref" : "#/definitions/Comment"
            }
          },
          "404" : {
            "description" : "Not found"
          },
          "500" : {
            "description" : "Unexpected error"
          }
        }
      },
      "put" : {
        "tags" : [ "Document comments" ],
        "summary" : "Modify a comment",
        "consumes" : [ "multipart/form-data" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "ID of the comment",
          "required" : true,
          "type" : "integer"
        }, {
          "name" : "description",
          "in" : "formData",
          "description" : "Content of the comment",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Successfully updated the comment",
            "schema" : {
              "$ref" : "#/definitions/Comment"
            }
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "404" : {
            "description" : "Not found"
          },
          "422" : {
            "description" : "Integrity error"
          },
          "500" : {
            "description" : "Unexpected error"
          }
        },
        "security" : [ {
          "Bearer" : [ ]
        } ]
      },
      "delete" : {
        "tags" : [ "Document comments" ],
        "summary" : "Delete a comment",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "ID of the comment",
          "required" : true,
          "type" : "integer"
        } ],
        "responses" : {
          "200" : {
            "description" : "Successfully deleted the comment",
            "schema" : {
              "$ref" : "#/definitions/Comment"
            }
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "404" : {
            "description" : "Not found"
          },
          "500" : {
            "description" : "Unexpected error"
          }
        }
      }
    },
    "/document/{id}/archive" : {
      "get" : {
        "tags" : [ "Document archives" ],
        "summary" : "Get all previous versions of a document with a specific ID",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "ID of the document",
          "required" : true,
          "type" : "integer"
        } ],
        "responses" : {
          "200" : {
            "description" : "The successfully retrieved versions",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/ExtendedDocument"
              }
            }
          },
          "404" : {
            "description" : "Not found"
          },
          "500" : {
            "description" : "Unexpected error"
          }
        }
      }
    },
...
  },
  "securityDefinitions" : {
    "Bearer" : {
      "type" : "apiKey",
      "name" : "Authorization",
      "in" : "header"
    }
  },
  "definitions" : {
    "TokenItem" : {
      "type" : "object",
      "properties" : {
        "token" : {
          "type" : "string",
          "format" : "jwt",
          "example" : "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJ1c2VyX2lkIjoxLCJ1c2VybmFtZSI6ImFkbWluIiwiZmlyc3ROYW1lIjoiQWRtaW4iLCJsYXN0TmFtZSI6IlVEViIsImVtYWlsIjoiYWRtaW5AdWR2LmZyIiwicG9zaXRpb24iOnsiTEVWRUxfTUlOIjoxLCJsYWJlbCI6ImFkbWluIiwiaWQiOjEsImNsZWFyYW5jZSI6WyJjb250cmlidXRvciIsInNvZnRNb2RlcmF0b3IiLCJtb2RlcmF0b3IiLCJhZG1pbiJdLCJMRVZFTF9NQVgiOjN9LCJleHAiOjE1NTg3NzcwMzYuMDM1MTczfQ.-6ns6_K-VpRlsqTrXbrV4shm8Oa0GJn3IfiR0LmLJoE"
        }
      }
    },
...
    "UserItem" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "example" : 1
        },
        "username" : {
          "type" : "string",
          "example" : "jdoe"
        },
        "firstName" : {
          "type" : "string",
          "example" : "John"
        },
        "lastName" : {
          "type" : "string",
          "example" : "Doe"
        },
        "email" : {
          "type" : "string",
          "example" : "john.doe@example.com"
        },
        "position" : {
          "$ref" : "#/definitions/PositionItem"
        },
        "position_id" : {
          "type" : "integer",
          "example" : 4
        }
      }
    },
...
    "ExtendedDocument" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "number",
          "example" : 1.0
        },
        "to_validate_doc" : {
          "type" : "number",
          "example" : 1.0
        },
        "user_id" : {
          "type" : "number",
          "example" : 1.0
        },
        "valid_doc" : {
          "$ref" : "#/definitions/ExtendedDocument_valid_doc"
        },
        "metaData" : {
          "$ref" : "#/definitions/MetaData"
        },
        "visualization" : {
          "$ref" : "#/definitions/Visualization"
        }
      }
    },
...
  },
...
}